+++ src/users/domain/dto/user.dto.ts
@@ -1,0 +1,11 @@
+export class UserDto {
+	readonly id: string
+
+	readonly userName: string
+
+	readonly email: string
+
+	createdAt: string
+
+	updatedAt: string
+}